requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dateutil==2.8.2
pandas==2.0.3
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.0.0
openai>=1.40.0
python-dotenv==1.0.0
Werkzeug==2.3.7
python-magic==0.4.27
python-magic-bin==0.4.14
